requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.2.0
pydantic[email]==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
stripe==6.0.0
httpx==0.27.0
requests==2.31.0
google-adk==0.0.9
